+++ include/hyperion/PriorityMuxer.h
@@ -54,6 +54,8 @@ public:
 		QString owner;
 	};
 
+	typedef QMap<int, InputInfo> InputsMap;
+
 	//Foreground and Background priorities
 	const static int FG_PRIORITY;
 	const static int BG_PRIORITY;
@@ -62,6 +64,9 @@ public:
 	const static int LOWEST_PRIORITY;
 	/// Timeout used to identify a non active priority
 	const static int TIMEOUT_NOT_ACTIVE_PRIO;
+	const static int REMOVE_CLEARED_PRIO;
+
+	const static int ENDLESS;
 
 	///
 	/// Constructs the PriorityMuxer for the given number of LEDs (used to switch to black when
@@ -164,7 +169,7 @@ public:
 	/// @param  timeout_ms  The new timeout (defaults to -1 endless)
 	/// @return             True on success, false when priority is not found
 	///
-	bool setInput(int priority, const std::vector<ColorRgb>& ledColors, int64_t timeout_ms = -1);
+	bool setInput(int priority, const std::vector<ColorRgb>& ledColors, int64_t timeout_ms = ENDLESS);
 
 	///
 	/// @brief   Update the current image of a priority (prev registered with registerInput())
@@ -173,7 +178,7 @@ public:
 	/// @param  timeout_ms  The new timeout (defaults to -1 endless)
 	/// @return             True on success, false when priority is not found
 	///
-	bool setInputImage(int priority, const Image<ColorRgb>& image, int64_t timeout_ms = -1);
+	bool setInputImage(int priority, const Image<ColorRgb>& image, int64_t timeout_ms = ENDLESS);
 
 	///
 	/// @brief Set the given priority to inactive
@@ -195,22 +200,13 @@ public:
 	///
 	void clearAll(bool forceClearAll=false);
 
-	///
-	/// @brief Queue a manual push where muxer doesn't recognize them (e.g. continuous single color pushes)
-	///
-	void queuePush() { emit timeRunner(); }
-
 signals:
-	///
-	/// @brief Signal which emits when a effect or color with timeout > -1 is running, once per second
-	///
-	void timeRunner();
 
 	///
 	/// @brief Emits whenever the visible priority has changed
 	/// @param  priority  The new visible priority
 	///
-	void visiblePriorityChanged(quint8 priority);
+	void visiblePriorityChanged(int priority);
 
 	///
 	/// @brief Emits whenever the current visible component changed
@@ -220,9 +216,13 @@ signals:
 
 	///
 	/// @brief Emits whenever something changes which influences the priorities listing
+
 	///        Emits also in 1s interval when a COLOR or EFFECT is running with a timeout > -1
+	/// @param  currentPriority The current priority at time of emit
+	/// @param  activeInputs The current active input map at time of emit
+
 	///
-	void prioritiesChanged();
+	void prioritiesChanged(int currentPriority, InputsMap activeInputs);
 
 	///
 	/// internal used signal to resolve treading issues with timer
@@ -231,15 +231,15 @@ signals:
 
 private slots:
 	///
-	/// Slot which is called to adapt to 1s interval for signal timeRunner() / prioritiesChanged()
+	/// Slot which is called to adapt to 1s interval for signal prioritiesChanged()
 	///
 	void timeTrigger();
 
 	///
-	/// Updates the current time. Channels with a configured time out will be checked and cleared if
-	/// required.
+	/// Updates the current priorities. Channels with a configured time out will be checked and cleared if
+	/// required. Cleared priorities will be removed.
 	///
-	void setCurrentTime();
+	void updatePriorities();
 
 private:
 	///
@@ -264,7 +264,7 @@ private:
 	hyperion::Components _prevVisComp = hyperion::COMP_INVALID;
 
 	/// The mapping from priority channel to led-information
-	QMap<int, InputInfo> _activeInputs;
+	InputsMap _activeInputs;
 
 	/// The information of the lowest priority channel
 	InputInfo _lowestPriorityInfo;