+++ libsrc/api/JsonAPI.cpp
@@ -185,12 +185,14 @@ proceed:
 		handleColorCommand(message, command, tan);
 	else if (command == "image")
 		handleImageCommand(message, command, tan);
+#if defined(ENABLE_EFFECTENGINE)
 	else if (command == "effect")
 		handleEffectCommand(message, command, tan);
 	else if (command == "create-effect")
 		handleCreateEffectCommand(message, command, tan);
 	else if (command == "delete-effect")
 		handleDeleteEffectCommand(message, command, tan);
+#endif
 	else if (command == "sysinfo")
 		handleSysInfoCommand(message, command, tan);
 	else if (command == "serverinfo")
@@ -275,6 +277,7 @@ void JsonAPI::handleImageCommand(const QJsonObject &message, const QString &comm
 	sendSuccessReply(command, tan);
 }
 
+#if defined(ENABLE_EFFECTENGINE)
 void JsonAPI::handleEffectCommand(const QJsonObject &message, const QString &command, int tan)
 {
 	emit forwardJsonMessage(message);
@@ -305,6 +308,7 @@ void JsonAPI::handleDeleteEffectCommand(const QJsonObject &message, const QStrin
 	const QString res = API::deleteEffect(message["name"].toString());
 	res.isEmpty() ? sendSuccessReply(command, tan) : sendErrorReply(res, command, tan);
 }
+#endif
 
 void JsonAPI::handleSysInfoCommand(const QJsonObject &, const QString &command, int tan)
 {
@@ -332,7 +336,9 @@ void JsonAPI::handleSysInfoCommand(const QJsonObject &, const QString &command, 
 	system["domainName"] = data.domainName;
 	system["isUserAdmin"] = data.isUserAdmin;
 	system["qtVersion"] = data.qtVersion;
+#if defined(ENABLE_EFFECTENGINE)
 	system["pyVersion"] = data.pyVersion;
+#endif
 	info["system"] = system;
 
 	QJsonObject hyperion;
@@ -359,20 +365,26 @@ void JsonAPI::handleServerInfoCommand(const QJsonObject &message, const QString 
 	QJsonArray priorities;
 	uint64_t now = QDateTime::currentMSecsSinceEpoch();
 	QList<int> activePriorities = _hyperion->getActivePriorities();
-	activePriorities.removeAll(255);
+	activePriorities.removeAll(PriorityMuxer::LOWEST_PRIORITY);
 	int currentPriority = _hyperion->getCurrentPriority();
 
-	for(int priority : activePriorities)
+	for(int priority : qAsConst(activePriorities))
 	{
 		const Hyperion::InputInfo &priorityInfo = _hyperion->getPriorityInfo(priority);
+
 		QJsonObject item;
 		item["priority"] = priority;
-		if (priorityInfo.timeoutTime_ms > 0)
+
+		if (priorityInfo.timeoutTime_ms > 0 )
+		{
 			item["duration_ms"] = int(priorityInfo.timeoutTime_ms - now);
+		}
 
 		// owner has optional informations to the component
 		if (!priorityInfo.owner.isEmpty())
+		{
 			item["owner"] = priorityInfo.owner;
+		}
 
 		item["componentId"] = QString(hyperion::componentToIdString(priorityInfo.componentId));
 		item["origin"] = priorityInfo.origin;
@@ -391,7 +403,8 @@ void JsonAPI::handleServerInfoCommand(const QJsonObject &message, const QString 
 			LEDcolor.insert("RGB", RGBValue);
 
 			uint16_t Hue;
-			float Saturation, Luminace;
+			float Saturation;
+			float Luminace;
 
 			// add HSL Value to Array
 			QJsonArray HSLValue;
@@ -485,6 +498,7 @@ void JsonAPI::handleServerInfoCommand(const QJsonObject &message, const QString 
 
 	info["adjustment"] = adjustmentArray;
 
+#if defined(ENABLE_EFFECTENGINE)
 	// collect effect info
 	QJsonArray effects;
 	const std::list<EffectDefinition> &effectsDefinitions = _hyperion->getEffects();
@@ -499,6 +513,7 @@ void JsonAPI::handleServerInfoCommand(const QJsonObject &message, const QString 
 	}
 
 	info["effects"] = effects;
+#endif
 
 	// get available led devices
 	QJsonObject ledDevices;
@@ -594,6 +609,10 @@ void JsonAPI::handleServerInfoCommand(const QJsonObject &message, const QString 
 
 #if defined(ENABLE_CEC)
 	services.append("cec");
+#endif
+
+#if defined(ENABLE_EFFECTENGINE)
+	services.append("effectengine");
 #endif
 
 #if defined(ENABLE_FORWARDER)
@@ -701,6 +720,7 @@ void JsonAPI::handleServerInfoCommand(const QJsonObject &message, const QString 
 	}
 	info["transform"] = transformArray;
 
+#if defined(ENABLE_EFFECTENGINE)
 	// ACTIVE EFFECT INFO
 	QJsonArray activeEffects;
 	for (const ActiveEffectDefinition &activeEffectDefinition : _hyperion->getActiveEffects())
@@ -717,6 +737,7 @@ void JsonAPI::handleServerInfoCommand(const QJsonObject &message, const QString 
 		}
 	}
 	info["activeEffects"] = activeEffects;
+#endif
 
 	// ACTIVE STATIC LED COLOR
 	QJsonArray activeLedColors;
@@ -1036,6 +1057,7 @@ void JsonAPI::handleSchemaGetCommand(const QJsonObject &message, const QString &
 	alldevices = LedDeviceWrapper::getLedDeviceSchemas();
 	properties.insert("alldevices", alldevices);
 
+#if defined(ENABLE_EFFECTENGINE)
 	// collect all available effect schemas
 	QJsonArray schemaList;
 	const std::list<EffectSchema>& effectsSchemas = _hyperion->getEffectSchemas();
@@ -1056,6 +1078,7 @@ void JsonAPI::handleSchemaGetCommand(const QJsonObject &message, const QString &
 		schemaList.append(schema);
 	}
 	properties.insert("effectSchemas", schemaList);
+#endif
 
 	schemaJson.insert("properties", properties);