+++ src/hyperiond/hyperiond.h
@@ -73,7 +73,9 @@ class JsonServer;
 class BonjourBrowserWrapper;
 class WebServer;
 class SettingsManager;
+#if defined(ENABLE_EFFECTENGINE)
 class PythonInit;
+#endif
 class SSDPHandler;
 class FlatBufferServer;
 class ProtoServer;
@@ -168,7 +170,9 @@ private:
 	AuthManager*               _authManager;
 	BonjourBrowserWrapper*     _bonjourBrowserWrapper;
 	NetOrigin*                 _netOrigin;
+#if defined(ENABLE_EFFECTENGINE)
 	PythonInit*                _pyInit;
+#endif
 	WebServer*                 _webserver;
 	WebServer*                 _sslWebserver;
 	JsonServer*                _jsonServer;